Solved 136. 8.7 seconds.
First solved this in Ruby in 124 seconds. Got the idea to just have
a pre-loaded prime number files from the PE forums. Then I implemented
it here.

I'm not checking in the prime number files for now because they are very
large. They can be recreated from files found on the interwebs.